backend/requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0